# src/main/java/io/nishandi/devopsapp/dto/EmployeeDto.java
package io.nishandi.devopsapp.dto;

import lombok.Getter;
import lombok.Setter;
import lombok.ToString;

@Getter
@Setter
@ToString
public class EmployeeDto {
    private String firstName;
    private String lastName;
}